Global standardization, AdamWR folder included, TPU acceleration added, LearningRateScheduler added, mixed precission capabilities added, etc.

+++ src/models/ImageBasedModels.py
@@ -30,7 +30,7 @@ class G2NetEfficientNet(object):
             self, 
             input_shape: Tuple[int, int, int],
             weights: str = "imagenet",
-            dtype: type = tf.float32
+            dtype: type = tf.float16
         ) -> None:
